requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pymupdf>=1.24.0,<2.0.0
pytesseract>=0.3.10,<1.0.0
pillow>=10.0.0,<12.0.0
pydantic>=2.7.0,<3.0.0
python-multipart>=0.0.9,<1.0.0
python-dotenv>=1.0.1,<2.0.0
openai>=1.40.0,<2.0.0
pytest>=8.0.0,<9.0.0
httpx>=0.27.0,<1.0.0
